Several Refactorings
Generified the search list, so now it is easier to add a multitude of different results to the search list.

Refactored the Person and Production panes out of the DetailPane class to increase module cohesion.

New method added to the SQL Server to get results for a person search query.